# docs/swagger.yaml
definitions:
...
  models.Bank:
    properties:
      created_at:
        type: string
      id:
        type: integer
      invar_coin:
        type: number
      tokens:
        items:
          $ref: '#/definitions/models.Token'
        type: array
      updated_at:
        type: string
      usdt_coin:
        type: number
      user_id:
        type: integer
    type: object
...
  models.Token:
    properties:
      bank_id:
        type: integer
      contract_address:
        type: string
      contract_chain:
        type: string
      contract_type:
        type: string
      created_at:
        type: string
      id:
        type: integer
      own_id:
        type: integer
      quantity:
        type: number
      updated_at:
        type: string
    type: object
...
  status.Response:
    properties:
      status_code:
        type: integer
      status_message:
        type: string
    type: object
  status.ResponseWtihData:
    properties:
      data:
        type: any
      status:
        $ref: '#/definitions/status.Response'
    type: object
info:
  contact:
    email: service@invar.finance
    name: InVar
    url: https://www.invar.finance/
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
  /admin/bank/{id}:
    get:
      consumes:
      - application/json
      description: 獲得使用者的財產狀況
      parameters:
      - description: Bank ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/status.ResponseWtihData'
            - properties:
                data:
                  $ref: '#/definitions/models.Bank'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.Response'
      security:
      - BearerAuth: []
      summary: 獲得使用者的財產狀況
      tags:
      - Bank
...
  /admin/change_password:
    post:
      consumes:
      - application/json
      description: 管理者修改密碼
      parameters:
      - description: 舊密碼
        in: body
        name: old_password
        required: true
        schema:
          type: string
      - description: 新密碼
        in: body
        name: password
        required: true
        schema:
          type: string
      - description: 確認新密碼
        in: body
        name: check_password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/status.ResponseWtihData'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.Response'
      security:
      - BearerAuth: []
      summary: 管理者修改密碼
      tags:
      - Auth
...
  /admin_auth:
    post:
      consumes:
      - application/json
      description: 後台管理者登入
      parameters:
      - description: 帳號
        in: body
        name: account
        required: true
        schema:
          type: string
      - description: 密碼
        in: body
        name: password
        required: true
        schema:
          type: string
      - description: 二階段驗證碼
        in: body
        name: tfa
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/status.ResponseWtihData'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.Response'
      summary: 後台管理者登入
      tags:
      - Auth
  /auth:
    post:
      consumes:
      - application/json
      description: 輸入帳號密碼與二階段驗證進行登入
      parameters:
      - description: 帳號
        in: body
        name: account
        required: true
        schema:
          type: string
      - description: 密碼
        in: body
        name: password
        required: true
        schema:
          type: string
      - description: 二階段驗證碼
        in: body
        name: tfa
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/status.ResponseWtihData'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.Response'
      summary: 使用者帳號登入
      tags:
      - Auth
  /bank:
    get:
      consumes:
      - application/json
      description: 獲得全部使用者的財產狀況
      parameters:
      - description: 信箱
        in: body
        name: email
        schema:
          type: string
      - description: 姓名
        in: body
        name: username
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/status.ResponseWtihData'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Bank'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.Response'
      security:
      - BearerAuth: []
      summary: 獲得全部使用者的財產狀況
      tags:
      - Bank
  /bank/{id}:
    get:
      consumes:
      - application/json
      description: 獲得使用者的財產狀況
      parameters:
      - description: Bank ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/status.ResponseWtihData'
            - properties:
                data:
                  $ref: '#/definitions/models.Bank'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.Response'
      security:
      - BearerAuth: []
      summary: 獲得使用者的財產狀況
      tags:
      - Bank
...
  /refresh_token:
    post:
      consumes:
      - application/json
      description: 更換令牌
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/status.ResponseWtihData'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.Response'
      summary: 更換令牌
      tags:
      - Auth
...
  /revoke_token:
    post:
      consumes:
      - application/json
      description: 撤銷令牌
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/status.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.Response'
      summary: 撤銷令牌
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"